fix(cloud): share the canonical router CORS origin policy
